perform unit conversions and list assumptions for calculations

+++ lidar/include/reading.hpp
@@ -5,8 +5,8 @@
 
 struct SensorReading final {
     public:
-        double distance;
-        double angle;
+        int distance;
+        int angle;
 };
 
 std::vector<SensorReading>
